refactor(view): rename components/@common to components/Common
