+++ packages/view/src/components/Common/GithubIssueLink/GithubIssueLink.scss
