+++ packages/view/src/components/Common/GithubIssueLink/renderIssueLinkedNodes.tsx
